Restructure Entity to Config
