+++ src/main/java/model/config/KeygroupMember.java
@@ -1,7 +1,5 @@
 package model.config;
 
-import model.Entity;
-
-public abstract class KeygroupMember extends Entity {
+public abstract class KeygroupMember extends Config {
 
 }
